backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic-extra-types>=2.0.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
pymongo>=4.0.0
imapclient>=3.0.0
python-multipart
python-jose[cryptography]>=3.3.0
python-decouple>=3.8